방 준비화면 ui (기능 적용)
