+++ src/pages/rooms/[roomId].tsx
@@ -14,6 +14,7 @@ import { getEnumKeyByEnumValue } from "@/utils/EnumUtil";
 import Button from "@mui/material/Button";
 import { RoomSettingDialog } from "@/components/RoomSettingDialog";
 import { useStores } from "@/stores/context";
+import WaitingRoom from "@/pages/waitingRoom";
 
 enum MasterPopupMenus {
   Kick = "강퇴",
@@ -22,6 +23,7 @@ enum MasterPopupMenus {
 
 const RoomScaffold: NextPage = observer(() => {
   const { userStore } = useStores();
+  const { roomListStore } = useStores();
   const [roomStore] = useState(() => new RoomStore(userStore));
   const router = useRouter();
   const roomId = router.query.roomId;
@@ -59,7 +61,7 @@ const NotExistsPage: NextPage = () => {
   );
 };
 
-const WaitingRoom: NextPage<{
+const WaitingRoom1: NextPage<{
   roomStore: RoomStore;
 }> = observer(({ roomStore }) => {
   return (
@@ -464,7 +466,7 @@ const DeviceSelector: NextPage<{ roomStore: RoomStore }> = observer(
   }
 );
 
-const Video: NextPage<{
+export const Video: NextPage<{
   id: string;
   videoStream: MediaStream | undefined;
   roomStore: RoomStore;
