+++ src/styles/enterRoom.module.scss
@@ -26,7 +26,7 @@
 
       .enter-room__user{
         width: 660px;
-        height: 720px;
+        height: 557px;
         background-color: #3D3D3D;
         display: flex;
         flex-direction: column;
